Rewording (again). Try to remove any implied msg
Alternative suggested wording:

"One delegate was loudly disruptive during a talk, and then obnoxious afterwards about how much they disagreed with it. Obviously it's fine to disagree with a speaker, but I expect delegates to respect our speakers during their talks, and if you want to share an opinion, I expect you to be open to receiving, listening to and respecting the validity of dissenting opinions from others. Simply being loudly obnoxious about promoting your own agenda is not acceptable behaviour, and this delegate won't be welcome at future events"